refactor([issue-746]): split autonomousJobs.js into focused submodules
The 1611-line autonomousJobs service became hard to navigate. Split it into a
server/services/autonomousJobs/ directory with one concern per file —
scriptHandlers, defaults, store (persistence/migration), scheduler, crud,
skillTemplates, and execution — and kept server/services/autonomousJobs.js as a
thin barrel re-exporting the exact same identifiers so every importer is
unchanged. No behavior change. Child-process/setTimeout error handling in the
shell/script execution paths is preserved verbatim.
